[FIX]  mail: prevent thread name in discuss ramh
Before this PR:

In the Discuss app, when you try to edit a thread name and leave the field empty
before pressing Enter, the name field stays empty.

After this PR

In the Discuss app, when you try to edit a thread name and leave the field empty
before pressing Enter, the previous name reappears.