[IMP] *: use new product name widget in purchase orders *: l10_in_purchase, purchase{,product_matrix}
This commit applies the new product name/description widget introduced
in https://github.com/odoo/odoo/commit/216634fd048c568d5b398134ba948a301bf76e3c on the purchase order lines in purchase order form view. This
combines both the name and description of the line in a single column
for better readability. It's also now not required to fill the
description field as before. The name and description of PO lines are
now also shown in the same way as the form view in report and portal
views.

Task-3930391