[IMP] hr_expense: show Create Report in mobile
Current behavior:

In mobile, 'Create Report' button is not shown in the list view, only
'Scan' and 'New' even if there are expenses to be reported. In Kanban
view, 'Create Report' is not shown at all.

Expected behavior:

'Create Report' should be shown in both List view and Kanban view, in
mobile or not, if there are expenses to be reported. To comply with
the existing style, in mobile:
- If there are expenses to reported, 'Scan' button is shown, with 'New'
and 'Create Report' as dropdown items.
- If there are no expenses to reported, 'Scan' and 'New' are displayed
(no dropdown).

This commit:

- Moves displayCreateReport and action_show_expenses_to_submit to
ExpenseDocumentUpload, so both Kanban and List can inherit them
- Creates a src/views/common.xml to store templates called by both List
and Kanban
- Changes display of button container class in mobile to block, to
ensure dropdown items show under each other instead of side-by-side

task-3981049