[IMP] l10n_in(_withholding): display alert message on absence of PAN
Under Section 206AA/206AB of the Income Tax Act, 1961, if the PAN of a
Vendor/Customer is unavailable, TDS/TCS should be at a higher rate.

With this commit, whenever a user applies a TDS/TCS tax and the PAN
number is absence of a partner, they will get an alert message to apply TDS/TCS
at a higher rate.

task-4037680