[IMP] utm: improve and generalize UTM trackers data
RATIONALE

The way we currently use the UTM trackers (aka source/medium/campaign) does not
comply with the industry standards, leading to confusion and globally unclear
or not usable reporting on UTM data.

PURPOSE

Clear the confusion of having specific social media as UTM medium.
The UTM medium should always be "Social Media" (general communication channel).

In addition, we also add the specific social media as UTM sources, for global
usage by other apps (notably Social).

To protect those common records and to prevent them to be deleted, we extend
the previous system to work for all utm models (source and medium notably).
Including making it less verbose and easier to use with "_utm_ref".

Task-3330036