[IMP] website_forum: duplicate forum tour for testing embed version
We duplicate the forum tour to test the embed version.

Technical note: we modify the step "This page contain new created question."
because it was actually testing an icon in the menu (the star in front of the
favorites menu rather than checking the post was displayed) which is not
present in the embed version.

Task-3974177