[FIX] web: `Dialog` header bg in the front-end
Before this commit: header color of `Dialog` on mobile was "community
color" or white, depending on the version of Odoo the database is
running. This was creating issues in in the front-end, where the user of
the website would see back-end specific colors in different `Dialog`s.

This commit removes the color customization of `Dialog`'header as it 
was fitting a design line we had before we introduced MILK.

task-4001365