[FIX] html_editor: make vertical seperators of toolbar have same length
Issue:
======
Vertical seperators doesn't have the same height in the toolbar.

Steps to reproduce the issue:
=============================
- Go to `to-do` app
- Add some text
- Select the text to show the toolbar
- Vertical seperators doesn't have the same height.

Reason behind the issue:
========================
The button groups doesnt have the same height and we use the
`border-left` to display the vertical seperator.

Solution:
=========
Use a span as a seperator so it won't depend anymore on the button group
styles.