[FIX] html_editor: adapt isolated link tests to new spec
This commit updates tests to the current spec, as they reflected the
editor's previous behaviour regarding links: upon click on a link, it
would be "isolated" for edition, by setting `contenteditable=true` to
the anchor element and `contenteditable=false` to the editable.

Such mechanism, that addresses shortcomings of contenteditable in regard
to links, has been replaced by the following one: the inner and outer
edges of links are now padded with Zero Width Non-Break Spaces, allowing
the user to place the cursor next to its content's edges, either inside
or outside the link, and insert text at the desired position.