[IMP] web: user switch
When several users share the same device, the login/logout process can
be boring, so we added a quick login feature.

For a security propose it doesn't store the session or the password of
the user. No data is sent to the server, everything is stored inside
the browser localstorage. It only stores the login of the user and
the avatar of the user for the last 5 users connected to the device.

task-3953707