[REF] maintenance: merge maintenance with project
Maintenance has been merged with project module such that:
1. `maintenance.team` --> `project.project`
2. `maintenance.request` --> `project.task`
3. `maintenance.stage` --> `project.task.type`

Each model has a boolean indicating if this is a "maintenance" object.
The way we are implementing the merge is like that after failing attempts
1. to inherit normally from project objects with different names.
Reason: a lot of many2many relationships needed to be handled separately
and the old views cannot be used as is
2. to inherit from project using delegation (composition)
Reason: ORM is currently not ready to handle this type of inheritance
relations, especially with so many computed fields.

task-3902350