[IMP] project: add "Close" button to project sharing wizard
Currently, when there are no collaborators in the project sharing
wizard and you click on the "Share Project" button, a toast will get
displayed indicating that the project was shared with the collaborators,
which is misleading.

This PR adds a "Close" button that is displayed instead of the "Share
Project" and "Discard" button when there are no collaborators.

task-4004946