[FIX] web: pushing the global state to the correct state
Changing commit title to push a runbot

Since [1], the global state was pushed before opening a new controller.
For instance, when opening a record in a kanban view. An issue could
happen, because the code didn't assure that the current state of the
router, was the same as the state of the current controller, before
pushing. The state of the router could be modified previously, for
instance, clicking on an internal link, so the state of the router will
be different as the state of the current controller.

This commit, will ensure that the pushed global state is done to the
correct state.