[FIX] website_slides_survey: prevent error while creating new certificate
The error we are encountering is due to the 'slide_id' key being missing in the
result dictionary. This is because the result dictionary contains an error
message instead of the expected data.

Steps to reproduce:
- Install the 'website_slides_survey' module(without demo data)
- Create a new 'Course' > 'Add Content' > Create new 'Certification'
- Add 'Title' and click on 'Save and Publish'

Traceback: KeyError
'slide_id'

This error occurs at [1] because the 'slide_id' is not present in the result
when adding a new certification without 'Select or create a certificate'.

This commit will fix the above error by displaying a warning indicating that
you must select a certificate in order to "Save and Publish" it.

[1]: https://github.com/odoo/odoo/blob/0e776e27e4e406e66853e7181889b3f54cd726e1/addons/website_slides_survey/controllers/slides.py#L80

sentry-5410807138