[IMP] mail: move thread open function to the correct bundle
The mail module is split into several sub-folders to determine when
the files should be loaded (e.g., web, public, public_web...).

Currently, several code snippets are misplaced, which increases the
load on public pages or live chat.

Moreover, this can lead to dependency hell, especially when moving a
feature to another bundle.

This PR fixes the `thread.open` method and its overrides.

Part of task-3972988