[REF] website_payment, *: convert s_donation options to OWL
*: web_editor
Given that the Donation options used to make a lot of modifications to
the DOM of the `we-list`, all of these have been converted into OWL-
compatible ways of modifying the `WeList` template.
Specifically, we cannot update the UserValueComponent's DOM on the fly
to add new inputs anymore, and we need a proper way to update the value
of the ListUserValue according to the option and vice-versa. Both of
these concerns are addressed by adapting the WeList's template and
adding a new prop `doubleInput`.

task-3850413