[IMP] pos_restaurant: imp order merge/split
Currently, the "Transfer/Merge" button allows the user to either
transfer the order to another table, or to link the table of the
current order with another table. When linking, the order of the old
table will be moved to the new table, but the 2 orders will not be
merged into one.

With the new changes, clicking on the button and then selecting a new
table will either perform a transfer, if the destination table does not
have an order on it, or an actual merging of the 2 orders into one, if
the destination table has an order. In this case, the tables will also
be automatically linked.

We also fix the issue that prevented floating orders to be transfered to
tables.

closes odoo/odoo#172513

Task: 3996948
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>