[FIX] html_editor: banner test fail due to spaces
The only difference between the expected and the actual result is the
presence of spaces in the expected result. This is due to the fact that
the spaces are actual text nodes with (invisible) spaces in the DOM,
which get removed when its parent element's has all of its content
deleted.