[FIX] html_editor: fix inline handling issues (non-editable, inline style)
Issue 1:
If a contenteditable='false' parent contains 2 nodes, one is
contenteditable='true' and the other is contenteditable='false', using
`deleteForward` or `deleteBackward` at the edges of the contenteditable='true'
should not remove the non-editable sibling (since it is itself part of a
non-editable).

Issue 2:
If the last inline element of a paragraph is contenteditable='false', and the
cursor is in at the edge of a sibling empty paragraph, using `deleteForward` or
`deleteBackward` should only remove the empty paragraph. Before this commit,
doing the delete did 2 removals at once: remove the empty paragraph and remove
the inline element. The correct behavior is to move the cursor next to the
inline non-editable element but not remove it (should be removed with a second
`delete` operation)

Issue 3:
If Flow content (not Phrasing content) was modified to have an `inline` display
style, it could be wrapped inside a paragraph during `initElementForEdition`.
which is invalid html. To avoid this issue, the functions wrapping inline
elements are putting them in a `div` for non-phrasing content (and still in a
`p` for phrasing content).