[IMP] website: add record fields as field visibility conditions
This commit adds support to select a single record as a form field
visibility condition.

The behavior is identical to visibility conditions on a select field,
except that the options are fetched (every record for the given model).

Usage:
Add the following attributes to the `s_website_form_field`:
 - `data-type="record"`
 - `data-model`: model name (i.e. `hr.job`)
 - `data-id-field` (optional, default=`id`): the model's primary key
   field name
 - `data-display-name-field` (optional, default=`display_name`): the
   name of the model field to use as a display string in select options

task-3864606