[REF] web_editor, *: convert ImageTools and WebsiteAnimate options
*: website, website_sale

This commit converts ImageTools and WebsiteAnimate options to OWL. For
website, a new WebsiteAnimateMixin is created to be able to combine both
sets of options into a single ImageToolsAnimate class in order to apply
hover effects on images: the previous code moved elements from one
widget to the other (back and forth) depending on some conditions, which
is neither sustainable nor advisable with the OWL framework.
As much as possible, the code from both ImageTools and WebsiteAnimate
related to hover effects is moved into ImageToolsAnimate.
Having all the methods in the same class also makes for a better and
easier communication between the different widgets.

For the same reason (avoid moving elements from one component to
another), the commit also adds a way for options to appear directly in
the Toolbar, to apply a TextAnimate class.

task-3850413