[FIX] core: actually test the condition we want to check
`odoo.http.request` always exists so making this import conditional is
nonsensical.

Since it's catching `RuntimeError`, this tries to check if the
`request` proxy is unbound: `AttributeError` is not a subtype of
`RuntimeError` so if the request has no env / cr / lang this crashes
anyway.

Checking if a local is bound can be done by just checking if it's
truthy.