[FIX] composer: composer assistant styling issue
When typing `=S` in the composer, the composer assistant will open.
If there wasn't enough space (default 300px width), this would break
the layout. This commit addresses the issue by adding a width check
for the composer assistant. If the width condition is met, the style
`right: 0px` is applied to prevent the layout from breaking.

Task: 4037213