[FIX] web, web_editor, *: store the correct image mimetype
*: website, website_event, website_sale

In some cases, `HTMLCanvasElement#toDataURL` returns a png instead of
the requested mimetype. Specifically, Safari doesn't support
`image/webp`. We call this an "implicit conversion".

Steps to reproduce (6 cases):
- Use Safari or an affected browser
- One of:
  - Image Field
    - In any regular page such as the Discuss home page, click on your
      user profile picture (top-right)
    - Click on Preferences
    - Change the profile picture to a webp image
  - Website
    - Insert a "Text - Image" snippet
    - Change the image format to one of the webp formats with a lower
      resolution
  - Website
    - Insert an "Image Gallery" snippet
    - Add a webp image
  - Website
    - Change the logo image
    - Select a webp format with a lower resolution than the original
  - Website/Events (website_event)
    - Change an event's cover image to a webp
  - Website/Shop (website_sale)
    - Change a product image's to a webp
- Save

Old behavior: the actual mimetype is `image/png` but it's saved as
`image/webp` and a .webp file extension
New behavior: it is correctly saved as `image/png` and a .png file
extension

To fix this, this commit handles all possible implicit conversions
through `convertCanvasToDataURL` and returns both the dataURL and the
actual mimetype in `applyModifications`. It then saves the output
mimetype and the correct file extension.

task-3847470