[FIX] web: pass context when duplicating records
i found that context is never passed in '_duplicateRecords' function for list views,
when making 'orm' calls from js side, while it is passed in 'duplicate' function of record,
which is called when duplicating record from form view.

use case:
In sale_renting module, There is difference in the result, when a sale_order is duplcated from
form view(duplicated properly) and when it is duplicated from list view(unexpected result).
When duplicating a sale_order, its lines are also being duplicated, and at that time '_compute_is_rental'
is being computed, which relies on context key 'in_rental_app'. 'in_rental_app' is set in the
related action yet was not available when duplicating the sale_order from list view (which resulted in unexpected behaviour).