[IMP] website: make link to plausible docs more explicit
The links pointing to the documentation have been refactored.
Previously, the link labeled "Connect Plausible" lacked explicit
clarity. We have now updated this link to follow a more recognizable
Odoo style, incorporating an arrow icon and a question mark to enhance
its visibility and intuitiveness.

For examples of similar links, refer to the settings section in Odoo.

task-3956742