[IMP] html_editor: selection normalization around contenteditable=false
Fully inclusion of partially selected contenteditable=false elements is
a requirement by the QWeb plugin, for its manages elements. Yet is was
implemented as the default selection normalization for any
contenteditable=false elements in the whole editor.

This commit applies removes such selection normalization and applies it
only for the concerned QWeb related elements.