[FIX] hr: prevent error when computing avatar_128 for resource
Steps to reproduce:
- Install 'planning' module
- Create a new user(eg: demo) and remove rights from 'Employee'
- Login with 'demo' in the new incognito window
- Now in the previous tab create one resource
- Go to the incognito window > Go to planning
- Click on 'New' > click on the 'Resource' field

Traceback: IndexError
tuple index out of range

The error occurs at [1] because we couldn't find an employee in
'avatar_per_employee_id'. This happens when a new resource is created in the
first tab, but the 'employee_id' is not found in the 'resource' in the second
tab.

This commit resolves the above error by returning false if an employee is not
present.

[1]- https://github.com/odoo/odoo/blob/fb4d758ed78211a61ea797759e1fb3b02b51be60/addons/hr/models/resource.py#L34

sentry-5508268100