[FIX] project_todo: basic adaptation for Breadcrumb
When navigating directly to a specific "To-Do", the custom breadcrumb
wasn't loaded properly, preventing the user from editing the to-do's
name, priority or state.

This commit, which initial aim was to make it "compatible" with the new
Breadcrumb's position on mobile (cf. in the Navbar instead of
ControlPanel), fixes this and the issue explained above by moving the
priority and state widgets to the CP's additional actions (aka. next to
the CogMenu) and moving the to-do's name into the stat buttons area,
next to the tags and assignees.

Also it allowed to enable the unit tests not only for desktop, but also
for mobile.

Note: as it is only a workable proof-of-concept and further refinement
are probably needed, the custom ControlPanel and editable breadcrumb
name files have not been deleted in this commit and related tests either
(only skipped).  Proper cleanup will be done later on when the said
refinements will take place.

task-3336242