[FIX] bus: reduce gevent worker concurrency
The gevent worker should be able to handle many databases. Currently,
several points reduce its efficiency:
- When no cursor is available to fetch bus notifications, websockets
retry up to 10 times to acquire one. However, the delay between each
attempt is too short, resulting in minimal chance of recovery when the
connection pool is full.
- The reconnect delay of the websocket worker is too short, causing many
websockets to try to reconnect in a short time frame when the server
restarts.
- The client updates the user presence every minute, which is too
frequent and causes unnecessary load on the server.

This PR fixes these issues by:
- Tweaking the `acquire_cursor` method: improving the  backoff and
increasing wait time to improve recovery chances.
- Increasing the presence update delay to 10 minutes, plus a random
jitter of up to 10 minutes.
- Increasing the websocket worker's initial reconnect delay to 1 second,
plus a random jitter of up to 30 seconds to spread reconnections over
a larger period.

opw-3896914,3874086