[IMP] web: open link in new tab with ctrl+click
This commit introduces the ability to open record in new tabs
by pressing ctrl key and then clicking on the record, or by using
the middle click on a mouse.

This behavior only applies on elements that don't have a native
experience handled by the browser (e.g. images/ urls).

task-4031939