[IMP] website: introduce `s_cta_box` snippet
This commit introduces a new inner snippet named `s_shape_image`.

Note: previous implementation allowed users to adapt the roundness of
the image as well as the width/height but this was dropped as considered
as too complex and the snippet was more fitting a two shapes scenario.

task-4094396
Part of task-4077427