[FIX] website_slides: fix statistics compute
During an upgrade some fields may be still to be loaded. Thus we cannot
blindly addume that the `bnr_...` fields are loaded yet. For example the
field `nbr_certification` from `website_slides_survey` module
https://github.com/odoo/odoo/blob/2d9f412a756776857c578708fd706c8d88fcf681/addons/website_slides_survey/models/slide_slide.py#L57
is not loaded at the time when `website_slides` is being loaded.

The root issue is that we are reading the category values direclty from
the DB via `_read_group` without filtering out the categories that are
already loaded.

Steps to reproduce the issue:
1. Install `website_slides_survey`
2. Try to re-install the module `website_slide` (`-i`)

We get the error:
```
  File "/home/odoo/src/odoo/15.0/addons/website_slides/models/slide_slide.py", line 347, in _compute_slides_statistics
    record.update(type_stats.get(record._origin.id, default_vals))
  File "/home/odoo/src/odoo/15.0/odoo/models.py", line 5665, in update
    record[name] = value
    ~~~~~~^^^^^^
  File "/home/odoo/src/odoo/15.0/odoo/models.py", line 5945, in __setitem__
    return self._fields[key].__set__(self, value)
           ~~~~~~~~~~~~^^^^^
KeyError: 'nbr_certification'
```