[FIX] website_blog: filters removal on date selection and removal
Steps to reproduce:
1. When we have multiple tag filters and if we add date filter then all
   other tags except first one are removed.
2. If there are more than one tag present in the filter section then
   only the first added tag filter remains and rest of tag filters are
   removed when try to remove the added date filter by clicking on X
   button present on this.

Expected Behaviour:
If anyone apply the date filter or remove the date filter then
previously added all the tags should also remain. No tags should be
removed after adding the date filter or removing the date filter.

This commit addresses the issue by sending a POST request whenever the
date filter is selected or removed. To achieve this, we introduced the
`post_link` class to the `<select>` and `<a>` elements.

task-3937884