[FIX] im_livechat, web: fix chat bot language on external website
When embedded on an external website, the live chat presents several
translation issues:
- The UI is translated according to the HTML lang attributes, but messages
posted afterwards (user joined/rating/chat bot) are not.
- Translations only work with an exact code match (e.g., `fr_FR` won't
fallback to `fr_BE`).

Live chat tries to guess the visitor's language at several points, each time
in a different fashion. This PR introduces a helper to ensure the language
is always computed in the same way:
- If the `frontend_lang` cookie is passed, use this language.
- If the user is logged in, use their language.
- Use the nearest language from the visitor's browser configuration.

opw-3997984