[FIX] account: fix fetch duplicate payment query
A check for duplicate payments was introduced in #155348. In the queries performed,
the outstanding_account_ids variable can be empty, causing the function
_fetch_duplicate_move_ids to crash.

To avoid that, = ANY() is used instead of IN, and part of the where condition is
added to the query as a parameter, conditional on whether outstanding_account_ids
is empty or not. Moreover, in case outstanding accounts are not found in the journal,
the company is also searched for outstanding accounts.

As IN tuple is not supported in psycopg3, IN is replaced by =ANY() in all instances
in the queries of _fetch_duplicate_move_ids, with the advantage that ANY accepts
an empty list as argument.

no-task