[IMP] product, website_sale: convert images to webp
Before this commit:
Previously, extra images and the main image for a product uploaded from the
product form view in the backend were not converted to `webp`, unlike those
uploaded from the frontend using the web editor.

After this commit:
Now, the main product image and the extra images will be converted to `webp`
when saving the product record.

task-3845144