[FIX] bus: reduce concurrency when fetching notifications
When no cursor is available to fetch bus notifications, websockets
retry up to 10 times to acquire one. However, the delay between each
attempt is too short, resulting in minimal chance of recovery when the
connection pool is full. This will eventually disconnect only to find
out no cursor is available to initialize the connection either.

This PR focuses on several points to ease recovery:
- Introduces a semaphore to reduce the likelihood of reaching this state
and increases the delay between each retry to allow the system to cool
down.
- Keeps 10% of the available cursors to handle incoming connection
requests.