[WIP] disable legacy options and keep editors in DOM
Prior to this, when dropping a snippets, its editors would be started
but not added to the DOM, which means that when clicking on it, there
would be a significant delay before options would be accessible.
Note that this delay is present when clicking on a Snippet that is not
dropped for the first time.

The reason for this delay is because even thought the widgets where not
in the DOM, they were still stored somewhere inside the started editor,
reducing the need for renders when clicking on a snippet that was
previously enabled.

This commit tries to mimic the old behaviour by keeping every element in
the DOM and mounted, just hidden with a d-none when the editor is not
currently enabled. This could have some side effects, like in tours
where a SnippetEditor selected is not the one currently visible.

For this reason, this commit is kept separate for now.